backend/fastapi_app/requirements.txt
# FastAPI Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database - PGVector
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Alternative: Pinecone (optional)
# pinecone-client>=2.2.4

# HTTP Client
httpx>=0.25.0

# JWT Authentication
python-jose[cryptography]>=3.3.0

# Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0


